+++ packages/expo-audio/src/AudioModule.web.ts
@@ -78,8 +78,8 @@ function getStatusFromMedia(media: HTMLMediaElement, id: number): AudioStatus {
   const status: AudioStatus = {
     id,
     isLoaded: true,
-    duration: media.duration * 1000,
-    currentTime: media.currentTime * 1000,
+    duration: media.duration,
+    currentTime: media.currentTime,
     playbackState: '',
     timeControlStatus: isPlaying ? 'playing' : 'paused',
     reasonForWaitingToPlay: '',
@@ -138,11 +138,11 @@ export class AudioPlayerWeb
   }
 
   get duration(): number {
-    return this.media.duration * 1000;
+    return this.media.duration;
   }
 
   get currentTime(): number {
-    return this.media.currentTime * 1000;
+    return this.media.currentTime;
   }
 
   get paused(): boolean {
@@ -189,7 +189,7 @@ export class AudioPlayerWeb
   }
 
   async seekTo(seconds: number): Promise<void> {
-    this.media.currentTime = seconds / 1000;
+    this.media.currentTime = seconds;
   }
 
   // Not supported on web