+++ packages/expo-sqlite/ios/SQLiteModuleLibSQL.swift
@@ -95,6 +95,7 @@ public final class SQLiteModule: Module {
         } else {
           // Try to find opened database for fast refresh
           if let cachedDb = findCachedDatabase(where: { $0.databasePath == databasePath && $0.openOptions == options && !options.useNewConnection }) {
+            cachedDb.addRef()
             return cachedDb
           }
 
@@ -158,12 +159,16 @@ public final class SQLiteModule: Module {
       }
 
       AsyncFunction("closeAsync") { (database: NativeDatabase) in
-        removeCachedDatabase(of: database)
-        try closeDatabase(database)
+        try maybeThrowForClosedDatabase(database)
+        if let db = removeCachedDatabase(of: database) {
+          try closeDatabase(db)
+        }
       }
       Function("closeSync") { (database: NativeDatabase) in
-        removeCachedDatabase(of: database)
-        try closeDatabase(database)
+        try maybeThrowForClosedDatabase(database)
+        if let db = removeCachedDatabase(of: database) {
+          try closeDatabase(db)
+        }
       }
 
       AsyncFunction("execAsync") { (database: NativeDatabase, source: String) in
@@ -492,7 +497,6 @@ public final class SQLiteModule: Module {
   }
 
   private func closeDatabase(_ db: NativeDatabase) throws {
-    try maybeThrowForClosedDatabase(db)
     for removedStatement in maybeRemoveAllCachedStatements(database: db) {
       if let rows = removedStatement.extraPointer {
         libsql_free_rows(rows)
@@ -698,9 +702,11 @@ public final class SQLiteModule: Module {
   private func removeCachedDatabase(of database: NativeDatabase) -> NativeDatabase? {
     return Self.lockQueue.sync {
       if let index = cachedDatabases.firstIndex(of: database) {
-        let database = cachedDatabases[index]
-        cachedDatabases.remove(at: index)
-        return database
+        let db = cachedDatabases[index]
+        if db.release() == 0 {
+          cachedDatabases.remove(at: index)
+          return db
+        }
       }
       return nil
     }