+++ docs/ui/components/Sidebar/SidebarSingleEntry.tsx
@@ -12,6 +12,7 @@ type SidebarSingleEntryProps = {
   isExternal?: boolean;
   secondary?: boolean;
   shouldLeakReferrer?: boolean;
+  allowCompactDisplay?: boolean;
 };
 
 export const SidebarSingleEntry = ({
@@ -21,7 +22,8 @@ export const SidebarSingleEntry = ({
   isActive = false,
   isExternal = false,
   secondary = false,
-  shouldLeakReferrer,
+  shouldLeakReferrer = false,
+  allowCompactDisplay = false,
 }: SidebarSingleEntryProps) => {
   return (
     <Tooltip.Root delayDuration={500} disableHoverableContent>
@@ -32,7 +34,7 @@ export const SidebarSingleEntry = ({
             'flex min-h-[32px] items-center gap-3 rounded-md px-2 py-1 text-sm !leading-[100%] text-secondary',
             'hocus:bg-element',
             'focus-visible:relative focus-visible:z-10',
-            'compact-height:justify-center compact-height:bg-subtle',
+            allowCompactDisplay && 'compact-height:justify-center compact-height:bg-subtle',
             secondary && 'text-xs',
             isActive &&
               '!bg-palette-blue3 font-medium text-link hocus:!bg-palette-blue4 hocus:text-link'
@@ -45,11 +47,15 @@ export const SidebarSingleEntry = ({
               isActive ? 'text-palette-blue11' : 'text-icon-tertiary'
             )}
           />
-          <span className="compact-height:hidden">{title}</span>
+          <span className={mergeClasses(allowCompactDisplay && 'compact-height:hidden')}>
+            {title}
+          </span>
           {isExternal && <ArrowUpRightIcon className="icon-sm ml-auto text-icon-secondary" />}
         </LinkBase>
       </Tooltip.Trigger>
-      <Tooltip.Content side="bottom" className="z-50 hidden compact-height:flex">
+      <Tooltip.Content
+        side="bottom"
+        className={mergeClasses('z-50 hidden', allowCompactDisplay && 'compact-height:flex')}>
         <span className="text-2xs text-secondary">{title}</span>
       </Tooltip.Content>
     </Tooltip.Root>