+++ packages/@expo/cli/src/start/server/metro/__tests__/fixtures/terminal-logs.ts
@@ -1,0 +1,11 @@
+// This was generated in SDK 52, before owner stack was introduced.
+// Created by rendering `<View>Hello</View>` in `expo/apps/router-e2e` on iOS in Expo Go.
+// I added a `console.log(JSON.stringify(event));` to MetroTerminalReporter._log.
+export const LOG_ERROR_TEXT_STRINGS_MUST_WRAPPED = {
+  type: 'client_log',
+  level: 'error',
+  data: [
+    'Warning: Text strings must be rendered within a <Text> component.\n    at RCTView (<anonymous>)\n    at View (http://192.168.1.245:8081/packages/expo-router/entry.bundle//&platform=ios&dev=true&hot=false&lazy=true&transform.engine=hermes&transform.bytecode=1&transform.routerRoot=__e2e__%2Fstatic-rendering%2Fapp&unstable_transformProfile=hermes-stable:62489:43)\n    at Page(./index.tsx) (<anonymous>)\n    at Suspense (<anonymous>)\n    at Route (http://192.168.1.245:8081/packages/expo-router/entry.bundle//&platform=ios&dev=true&hot=false&lazy=true&transform.engine=hermes&transform.bytecode=1&transform.routerRoot=__e2e__%2Fstatic-rendering%2Fapp&unstable_transformProfile=hermes-stable:122585:24)\n    at RCTView (<anonymous>)\n    at View (http://192.168.1.245:8081/packages/expo-router/entry.bundle//&platform=ios&dev=true&hot=false&lazy=true&transform.engine=hermes&transform.bytecode=1&transform.routerRoot=__e2e__%2Fstatic-rendering%2Fapp&unstable_transformProfile=hermes-stable:62489:43)\n    at AppContainer (http://192.168.1.245:8081/packages/expo-router/entry.bundle//&platform=ios&dev=true&hot=false&lazy=true&transform.engine=hermes&transform.bytecode=1&transform.routerRoot=__e2e__%2Fstatic-rendering%2Fapp&unstable_transformProfile=hermes-stable:62357:25)\n    at main(RootComponent) (http://192.168.1.245:8081/packages/expo-router/entry.bundle//&platform=ios&dev=true&hot=false&lazy=true&transform.engine=hermes&transform.bytecode=1&transform.routerRoot=__e2e__%2Fstatic-rendering%2Fapp&unstable_transformProfile=hermes-stable:85226:28)',
+  ],
+  mode: 'NOBRIDGE',
+} as const;