+++ packages/expo-local-authentication/src/LocalAuthentication.ts
@@ -8,6 +8,7 @@ import {
   LocalAuthenticationResult,
   SecurityLevel,
   BiometricsSecurityLevel,
+  LocalAuthenticationError,
 } from './LocalAuthentication.types';
 
 export {
@@ -16,6 +17,7 @@ export {
   LocalAuthenticationResult,
   SecurityLevel,
   BiometricsSecurityLevel,
+  LocalAuthenticationError,
 };
 
 // @needsAudit
@@ -36,7 +38,7 @@ export async function hasHardwareAsync(): Promise<boolean> {
  * Determine what kinds of authentications are available on the device.
  * @return Returns a promise which fulfils to an array containing [`AuthenticationType`s](#authenticationtype).
  *
- * Devices can support multiple authentication methods- i.e. `[1,2]` means the device supports both
+ * Devices can support multiple authentication methods - i.e. `[1,2]` means the device supports both
  * fingerprint and facial recognition. If none are supported, this method returns an empty array.
  */
 export async function supportedAuthenticationTypesAsync(): Promise<AuthenticationType[]> {