+++ packages/expo-modules-autolinking/android/expo-gradle-plugin/expo-autolinking-settings-plugin/src/main/kotlin/expo/modules/plugin/SettingsManager.kt
@@ -2,7 +2,6 @@ package expo.modules.plugin
 
 import expo.modules.plugin.configuration.ExpoAutolinkingConfig
 import expo.modules.plugin.configuration.GradleProject
-import expo.modules.plugin.configuration.MavenRepo
 import expo.modules.plugin.gradle.afterAndroidApplicationProject
 import expo.modules.plugin.gradle.applyAarProject
 import expo.modules.plugin.gradle.applyPlugin
@@ -10,6 +9,7 @@ import expo.modules.plugin.gradle.beforeProject
 import expo.modules.plugin.gradle.beforeRootProject
 import expo.modules.plugin.gradle.linkAarProject
 import expo.modules.plugin.gradle.linkBuildDependence
+import expo.modules.plugin.gradle.linkLocalMavenRepository
 import expo.modules.plugin.gradle.linkMavenRepository
 import expo.modules.plugin.gradle.linkPlugin
 import expo.modules.plugin.gradle.linkProject
@@ -120,16 +120,22 @@ class SettingsManager(
 
       // Adds maven repositories for all projects that are using the publication.
       // It most likely means that we will add "https://maven.pkg.github.com/expo/expo" to the repositories.
-      config
+      val localRepositories = config
         .allProjects
         .filter { it.usePublication && it.publication?.repository != "mavenLocal" }
         .mapNotNull {
-          it.publication?.repository
+          val publication = it.publication
+            ?: return@mapNotNull null
+
+          "${it.sourceDir}/../${publication.repository}" to publication
         }
-        .toSet()
-        .forEach { url ->
-          rootProject.linkMavenRepository(MavenRepo(url))
+        .groupBy({ it.first }, { it.second })
+
+      rootProject.allprojects { project ->
+        localRepositories.forEach { (path, publications) ->
+          project.linkLocalMavenRepository(path, publications)
         }
+      }
     }
 
     settings.gradle.afterAndroidApplicationProject { androidApplication ->