+++ packages/expo-ui/ios/PickerView.swift
@@ -12,11 +12,15 @@ class PickerProps: ExpoSwiftUI.ViewProps {
   var onOptionSelected = EventDispatcher()
 }
 
-struct PickerView: ExpoSwiftUI.View {
+struct PickerView: ExpoSwiftUI.View, ExpoSwiftUI.WithHostingView {
   @State var selection: Int = 0
   @State var prevSelectedIndex: Int?
-  @EnvironmentObject var props: PickerProps
+  @ObservedObject var props: PickerProps
   @EnvironmentObject var shadowNodeProxy: ExpoSwiftUI.ShadowNodeProxy
+
+  init(props: PickerProps) {
+    self.props = props
+  }
 
   var body: some View {
     if #available(iOS 17.0, tvOS 17.0, *) {