+++ packages/expo-ui/ios/SectionView.swift
@@ -10,8 +10,8 @@ class SectionProps: ExpoSwiftUI.ViewProps {
 let IPAD_OFFSET: CGFloat = 30
 let IPHONE_OFFSET: CGFloat = 40
 
-struct SectionView: ExpoSwiftUI.View {
-  @EnvironmentObject var props: SectionProps
+struct SectionView: ExpoSwiftUI.View, ExpoSwiftUI.WithHostingView {
+  @ObservedObject var props: SectionProps
 
   var body: some View {
     let form = Form {